Centralise the logic for combining metadata and body writers
